Extract Progression interface
